// src/types/web-page-test.ts
/**
 * @license
 * Copyright 2019 Google Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Cdn } from './cdn';
import { Connectivity } from './connectivity';

export const enum TestStatus {
  Succeeded,
  Failed,
  Pending,
}

interface TestPage {
  path: string;
}

export interface NetworkLatencyTestPage extends TestPage {
  cdn: Cdn;
}

export interface ExecutionLatencyTestPage extends TestPage {}

export const enum TestType {
  NetworkLatency = 'network_latency',
  ExecutionLatency = 'execution_latency',
}

export interface TestRecord {
  status: TestStatus;
  url: string;
  device: string;
  connectivity: Connectivity;
  runs: number;
  start: number;
  timeout: number;
}